@@ -3,6 +3,7 @@ import {Component} from '@angular/core';
 import {Platform} from '@ionic/angular';
 import {SplashScreen} from '@ionic-native/splash-screen/ngx';
 import {StatusBar} from '@ionic-native/status-bar/ngx';
+import {AuthService} from './services/auth.service';
 
 @Component({
     selector: 'app-root',
@@ -13,7 +14,8 @@ export class AppComponent {
     constructor(
         private platform: Platform,
         private splashScreen: SplashScreen,
-        private statusBar: StatusBar) {
+        private statusBar: StatusBar,
+        private authService: AuthService) {
         this.initializeApp();
     }
 
@@ -22,5 +24,7 @@ export class AppComponent {
             this.statusBar.styleDefault();
             this.splashScreen.hide();
         });
+
     }
+
 }
