@@ -1,7 +1,6 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {FormsModule} from '@angular/forms';
-
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {IonicModule} from '@ionic/angular';
 
 import {BookingInfoPageRoutingModule} from './booking-info-routing.module';
@@ -13,13 +12,16 @@ import {StepOneComponent} from './step-one/step-one.component';
 import {StepTwoComponent} from './step-two/step-two.component';
 import {PaymentCardComponent} from './payment-card/payment-card.component';
 
+
 @NgModule({
     imports: [
         CommonModule,
         FormsModule,
+        ReactiveFormsModule,
         IonicModule,
         BookingInfoPageRoutingModule,
-        ComponentModule
+        ComponentModule,
+        // Stripe
     ],
     declarations: [
         BookingInfoPage,
