@@ -12,7 +12,9 @@ export class BookingInfoPage implements OnInit {
     textVn: any = {
         step1Text: 'Thông tin đặt chỗ',
         step2Text: 'Xác nhận và thanh toán',
-        bookingInfoBtn: 'Xác nhận thông tin thanh toán'
+        bookingInfoBtn: 'Xác nhận thông tin thanh toán',
+        title: 'Thanh toán bằng thẻ quốc tế Visa, Master, JCB',
+        description: 'Nhập thông tin hoặc lựa chọn thẻ đã lưu để tiến hành thanh toán quốc tế qua Stripe.'
     };
     textEn: any = {
         step1Text: 'Booking information',
@@ -22,18 +24,22 @@ export class BookingInfoPage implements OnInit {
     text: any = {};
     step: number;
     headerStyle: any = HeaderStyle;
+
+    isMobile = window.innerWidth < 767;
+
     constructor(private route: ActivatedRoute,
                 private router: Router) {
         this.text = this.lang === 'en' ? this.textEn : this.textVn;
     }
 
+
     ngOnInit() {
         this.route.queryParamMap.subscribe((params: ParamMap) => {
-           if (!params.has('step')) {
-               this.router.navigate(['/']);
-           } else {
-               this.step = +params.get('step') as number;
-           }
+            if (!params.has('step') || !params.has('booking_id')) {
+                this.router.navigate(['/']);
+            } else {
+                this.step = +params.get('step') as number;
+            }
         });
     }
 
