@@ -1,4 +1,7 @@
 import {Component, OnInit} from '@angular/core';
+import {Booking, BookingService} from '../../../../../swagger';
+import {ActivatedRoute, Router} from '@angular/router';
+import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 
 @Component({
     selector: 'app-step-one',
@@ -7,9 +10,20 @@ import {Component, OnInit} from '@angular/core';
 })
 export class StepOneComponent implements OnInit {
     lang = localStorage.getItem('lang');
+
+    booking: Booking = {};
+    bookingId = this.route.snapshot.queryParams.booking_id;
+    bookingForm: FormGroup = this.fb.group({
+        customerName: this.fb.control(null, Validators.required),
+        totalPrice: this.fb.control(null, Validators.required),
+        guestCount: this.fb.control(null, Validators.required),
+        phone: this.fb.control(null, Validators.required),
+        email: this.fb.control(null, [Validators.required, Validators.email]),
+    });
+
     textVn: any = {
         step1Text: 'Thông tin đặt chỗ',
-        step12Text: 'THÔNG TIN CÁ NHÂN',
+        step12Text: 'Thông tin cá nhân',
         step13Text: '2 đêm tại S. CAMELLIA 2 _ Managed by SONG CAT HOME',
         numCustomers: 'Số khách',
         customerName: 'Tên khách hàng',
@@ -18,7 +32,8 @@ export class StepOneComponent implements OnInit {
         startDateBookinng: 'Ngày nhận phòng',
         endDateBookinng: 'Ngày trả phòng',
         step1TextDescription: 'Kiểm tra lại thông đặt phòng của quý khách hàng',
-        bookingInfoBtn: 'Xác nhận thông tin thanh toán'
+        bookingInfoBtn: 'Xác nhận thông tin thanh toán',
+        totalPrice: 'Tổng thanh toán'
     };
     textEn: any = {
         step1Text: 'Booking information',
@@ -31,15 +46,59 @@ export class StepOneComponent implements OnInit {
         startDateBookinng: 'Start day booking',
         endDateBookinng: 'End day booking',
         step1TextDescription: 'Check information of your booking',
-        bookingInfoBtn: 'Booking info confirm'
+        bookingInfoBtn: 'Booking info confirm',
+        totalPrice: 'Total price'
     };
     text: any = {};
 
-    constructor() {
+    constructor(private router: Router,
+                private fb: FormBuilder,
+                private route: ActivatedRoute,
+                private bookingService: BookingService) {
         this.text = this.lang === 'en' ? this.textEn : this.textVn;
     }
 
     ngOnInit() {
+        this.bookingService.getBookingById(this.bookingId)
+            .subscribe(
+                (booking: Booking) => {
+                    this.booking = booking;
+                    this.updateBookingForm();
+                },
+                () => {
+                    this.router.navigate(['/pages']);
+                }
+            );
     }
 
+    updateBookingForm() {
+        Object.keys(this.booking).forEach(key => {
+            if (this.bookingForm.get(key)) {
+                this.bookingForm.get(key).reset(this.booking[key]);
+            }
+        });
+    }
+
+    updateBookingThenStep2() {
+        const bookingUpdate: Booking = {
+            ...this.booking,
+            ...this.bookingForm.value,
+        };
+
+        this.bookingService.updateBooking(bookingUpdate)
+            .subscribe((booking) => {
+                this.goToStep2();
+            });
+    }
+
+    goToStep2() {
+        this.router.navigate(
+            [],
+            {
+                relativeTo: this.route,
+                queryParams: {step: 2},
+                queryParamsHandling: 'merge'
+            }
+        );
+    }
 }
