@@ -1,24 +1,26 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 
-import { IonicModule } from '@ionic/angular';
+import {IonicModule} from '@ionic/angular';
 
-import { LoginPageRoutingModule } from './login-routing.module';
+import {LoginPageRoutingModule} from './login-routing.module';
 
-import { LoginPage } from './login.page';
+import {LoginPage} from './login.page';
 import {GoFormLoginComponent} from './form-login/form-login.component';
 
 @NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-    LoginPageRoutingModule
-  ],
-  declarations: [
-      LoginPage,
-      GoFormLoginComponent
-  ]
+    imports: [
+        CommonModule,
+        FormsModule,
+        IonicModule,
+        LoginPageRoutingModule,
+        ReactiveFormsModule
+    ],
+    declarations: [
+        LoginPage,
+        GoFormLoginComponent
+    ]
 })
-export class LoginPageModule {}
+export class LoginPageModule {
+}
