@@ -4,7 +4,9 @@
 
 export const environment = {
   production: false,
-  basePath: 'https://virtserver.swaggerhub.com/unhang/WannaGo/1.0.0'
+  // basePath: 'https://virtserver.swaggerhub.com/unhang/WannaGo/1.0.0',
+  basePath: 'https://wannago.cf',
+  // basePath: 'https://wannago.ml'
 };
 
 /*
